Code refactoring, settings are finally functional
