+++ Procfile
@@ -1,1 +1,1 @@
-worker: java $JAVA_OPTS -cp build/staging:build/staging/* org.qweco.vocabularybot.Bot
+worker: java $JAVA_OPTS -cp build/staging:build/staging/* org.qweco.vocabularybot.TelegramBot
